Refactor disk structs into disks/structs
