+++ astoria/common/disks/__init__.py
@@ -1,1 +1,9 @@
 """Code for recognising disk drives."""
+
+from .structs import DiskInfo, DiskType, DiskUUID
+
+__all__ = [
+    "DiskInfo",
+    "DiskType",
+    "DiskUUID",
+]
