+++ astoria/managers/astdiskd/disk_manager.py
@@ -4,12 +4,9 @@ import logging
 from pathlib import Path
 from typing import List
 
+from astoria.common.disks import DiskInfo, DiskType
 from astoria.common.manager import StateManager
-from astoria.common.messages.astdiskd import (
-    DiskInfo,
-    DiskManagerMessage,
-    DiskType,
-)
+from astoria.common.messages.astdiskd import DiskManagerMessage
 
 from .disk_provider import DiskProvider
 from .static import StaticDiskProvider
